GFGs POTD Solution 12/12/2023
This is the first Dynamic Programming Problem I have solved completely without any help. It is a medium level problem at GFGs platform. It took me around 1 hour to completely solve but still it gave a boost to my confidence. 

Thanks to Mr. Kamal Rawat for teaching a clear and better approach for DP Problems.